优化了布局，增加了修改任务和删除任务功能

+++ app/src/main/java/com/levent_j/potato/utils/Util.java
@@ -4,8 +4,8 @@ package com.levent_j.potato.utils;
  * Created by levent_j on 16-5-20.
  */
 public class Util {
-    public static double Minute2Second(double minute){
+    public static int Minute2Second(double minute){
         //*60000
-        return minute*5000;
+        return (int)minute*5000;
     }
 }
